fix - step2 - 코드리뷰 반영
