--- src/main/kotlin/step2/StringCalculator.kt
@@ -1,10 +1,0 @@
-package step2
-
-class StringCalculator {
-    companion object {
-        fun execute(input: String?): Int {
-            InputValidator.validate(input)
-            return Calculator.calculate(input!!)
-        }
-    }
-}
