+++ src/main/kotlin/step2/calculator/StringCalculator.kt
@@ -1,0 +1,8 @@
+package step2.calculator
+
+object StringCalculator {
+    fun execute(input: String?): Int {
+        InputValidator.validate(input)
+        return Calculator.calculate(input!!)
+    }
+}
